refactor: move logic to custom hook
