+++ src/pages/oAuth/github/ui/GitHubPage.tsx
@@ -1,31 +1,9 @@
 'use client'
 
-import { useRouter, useSearchParams } from 'next/navigation'
-import { useDispatch } from 'react-redux'
-import { ROUTES } from '@/shared/router/routes'
-import { useEffect } from 'react'
-import { tokenReceived, useGithubUpdateTokensMutation } from '@/entities/user'
+import { useGithubPage } from '@/pages/oAuth/github/lib/useGithubPage'
 
 export function GithubPage() {
-  const router = useRouter()
-  const searchParams = useSearchParams()
-  const dispatch = useDispatch()
-  const accessToken = searchParams!.get('accessToken')
-  const email = searchParams!.get('email')
-  const [githubUpdateTokens, { isSuccess }] = useGithubUpdateTokensMutation()
-
-  useEffect(() => {
-    if (accessToken && email) {
-      dispatch(tokenReceived({ accessToken }))
-      githubUpdateTokens()
-    }
-  }, [accessToken, dispatch, email, githubUpdateTokens])
-
-  useEffect(() => {
-    if (isSuccess) {
-      router.replace(ROUTES.HOME)
-    }
-  }, [isSuccess, router])
+  const { accessToken, email } = useGithubPage()
 
   if (!accessToken && !email) {
     return null
